refactor(core): extract NCD to node-free core/ncd.ts
Move ncd() + its gzip helper out of proofs.ts (which imports node:crypto via
hash.js) into a node-free leaf so the description-overlap detector reaches it
without dragging crypto into the in-browser audit bundle. proofs.ts re-exports
ncd for its existing consumers; description-overlap.ts imports it from the leaf.
Swaps Buffer.from(s,'utf-8') → TextEncoder so the byte input is identical in
Node and the browser and the node-only Buffer global is gone (the browser build
aliases node:zlib → pako). No behaviour change — parity + proofs + overlap
tests green.
